Take the logarithm of the absolute value of a number with its sign

This function takes a numerical input x, computes the logarithm of its absolute value and multiplies it by the sign of x. This ensures that the result has the same sign as the input x.

Usage

.logmod(x)

Arguments

x

A numerical input.

Value

The result of taking the logarithm of the absolute value of x with its sign.

Examples

.logmod(10)
#> [1] 2.302585
.logmod(-5)
#> [1] -1.609438