Plot trinary ROC curve and save the output as pdf file.

Plot trinary ROC curve and save the output as pdf file.

Usage

trinaryROCPlot(trinaryPlotThings, trinaryDF, plotFile = NULL, openFig = TRUE)

Arguments

trinaryPlotThings

Object of trinary ROC curve, contains xx, y, xout, and y. as its attributes.

trinaryDF

Data frame with trinary ROC curve information, including youden.thresh.roc.x, hi.thresh.roc.x, lo.thresh.roc.x, and trinary.pauc.

plotFile

Path to save the output as pdf file. If NULL, the plot will not be saved.

openFig

Logical, if TRUE, the plot will be shown in a new window.

Value

None